# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RoomFinder",
        "description": "RoomFinder API description",
        "contact": {
            "email": "manis6062@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/jagga/add": {
            "post": {
                "summary": "Add Jagga",
                "operationId": "addJagga",
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "header",
                        "description": "Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User Id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "Type",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone_no",
                        "in": "formData",
                        "description": "Phone no.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "loc_lat",
                        "in": "formData",
                        "description": " Location - Longitude",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "loc_lon",
                        "in": "formData",
                        "description": "Location - Latitude",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "description": "Address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "image[]",
                        "in": "formData",
                        "description": "Jagga Image",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "Price",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sold",
                        "in": "formData",
                        "description": "(0 for unsold or 1 for sold)",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/jagga/search-jagga": {
            "get": {
                "summary": "Search Jagga",
                "operationId": "searchJagga",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User Id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "Type",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone_no",
                        "in": "formData",
                        "description": "Phone no.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "loc_lat",
                        "in": "formData",
                        "description": " Location - Longitude",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "loc_lon",
                        "in": "formData",
                        "description": "Location - Latitude",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "description": "Address",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "Room Image",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "Price",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "sold",
                        "in": "formData",
                        "description": "(0 for unsold or 1 for sold)",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/jagga/my-favourite-jaggas": {
            "get": {
                "summary": "My Favourite jaggas",
                "operationId": "myFavouriteJaggas",
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "header",
                        "description": "Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "per_page",
                        "in": "formData",
                        "description": "Jaggas Per Page",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page_number",
                        "in": "formData",
                        "description": "Per Page Number",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/jagga/delete": {
            "delete": {
                "summary": "Delete Jagga",
                "operationId": "deleteJagga",
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "header",
                        "description": "Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "User Id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "jagga_id",
                        "in": "formData",
                        "description": "Jagga Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}